Modificaciones menores a la clase Profesor

+++ src/Entities/Profesor.java
@@ -18,14 +18,14 @@ public class Profesor
     private String nombre;
     private String apellido;
     private String despacho;
-    private int telefono;
-    private int participaciones;
+    private float telefono;
+    private float participaciones;
     private String GradEstudios;
 
     public Profesor() {
     }
 
-    public Profesor(String nombre, String apellido, String despacho, int telefono, int participaciones, String GradEstudios) {
+    public Profesor(String nombre, String apellido, String despacho, float telefono, float participaciones, String GradEstudios) {
         this.nombre = nombre;
         this.apellido = apellido;
         this.despacho = despacho;
@@ -34,7 +34,7 @@ public class Profesor
         this.GradEstudios = GradEstudios;
     }
 
-    public Profesor(ObjectId id, String nombre, String apellido, String despacho, int telefono, int participaciones, String GradEstudios) {
+    public Profesor(ObjectId id, String nombre, String apellido, String despacho, float telefono, float participaciones, String GradEstudios) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -84,23 +84,19 @@ public class Profesor
         this.despacho = despacho;
     }
 
-    public int getTelefono()
-    {
+    public float getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(int telefono)
-    {
+    public void setTelefono(float telefono) {
         this.telefono = telefono;
     }
 
-    public int getParticipaciones() 
-    {
+    public float getParticipaciones() {
         return participaciones;
     }
 
-    public void setParticipaciones(int participaciones)
-    {
+    public void setParticipaciones(float participaciones) {
         this.participaciones = participaciones;
     }
